+++ protocols/RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 # dxFileAccessProtocols
 
-## in develop
+## 0.2.0 (2021-05-07)
 
 * Fixes handling of `DxFolderSource` for root folders
 * Implements `container` and `version` fields for dx and S3 protocols